requirements.txt
django==3.1.14
django-cors-headers==3.2.0
django-picklefield==2.0
django-request-logging==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.11.2
pytest-django==4.4.0
